Breathly v2 (#99)
I redesigned Breathly's UI and architecture from the ground up to make it easy to implement many of the requests I got in the past. I'm talking about things like pure OLED theme, being able to use decimals in the exercise step lengths, adding white noise, and so on.

From a technical standpoint, I moved Breathly to Expo. This change should make it easier to update to major versions of React Native.